Start looking at doing checkpointing
Log some checkpoint information into a file.

We log the playlist/uri being played, the position in the playlist
and the time elapsed of the current track in nano seconds.

This data is logged into a temporary file, then rename()'d to it's
final place. This should ensure that we don't get a corrupted
checkpoint file.

The use of fsync/fdatasync/sync_file_range should also be investigated.

This is just a first stab at seeing what's involved and what information
is required to be logged. It is as such likely to change somewhat.